requirements.txt
aiofiles==23.2.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
babel==2.17.0
backrefs==5.9
bcrypt==4.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.104.1
flower==2.0.1
ghp-import==2.1.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
humanize==4.13.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.20
lightgbm==4.6.0
limits==5.5.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.18
mkdocs-material-extensions==1.3.1
numpy==2.3.2
nvidia-nccl-cu12==2.27.7
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
paginate==0.5.7
pandas==2.3.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymdown-extensions==10.16.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
redis==4.6.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==1.38.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
structlog==25.4.0
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.3
xgboost==3.0.4
xlrd==2.0.2
xxhash==3.5.0
zstandard==0.24.0